requirements.txt
langchain
wikipedia
langchain-community
python-dotenv
pydantic
langchain-google-genai
duckduckgo-search
streamlit